+++ molo/profiles/migrations/0015_add_curent_site_to_user_profile.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.db import migrations, models
+from django.db import migrations
 
 
 def add_site_to_user_profile(apps, schema_editor):
@@ -11,11 +11,12 @@ def add_site_to_user_profile(apps, schema_editor):
     if site:
         UserProfile.objects.filter(user__is_staff=False).update(site=site)
 
+
 class Migration(migrations.Migration):
 
     dependencies = [
         ('wagtailcore', '__latest__'),
-        ('profiles', '0013_add_site_to_profile'),
+        ('profiles', '0014_add_site_to_profile'),
     ]
 
     operations = [