+++ molo/profiles/models.py
@@ -85,6 +85,97 @@ class UserProfilesSettings(BaseSetting):
         related_name='+',
         help_text=_('Choose a footer page')
     )
+    activate_display_name = models.BooleanField(
+        default=True,
+        editable=True,
+        verbose_name=_("Activate Display Name"),
+    )
+    capture_display_name_on_reg = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Capture On Registration"),
+        help_text=_("If Display Name is activated, "
+                    "and Capture On Registration is not activated, "
+                    "The display name field will be captured on done page."),
+    )
+    display_name_required = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Display Name Required"),
+    )
+    activate_gender = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Activate Gender"),
+    )
+    capture_gender_on_reg = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Capture On Registration"),
+        help_text=_("If Gender is activated, "
+                    "and Capture On Registration is not activated, "
+                    "The Gender field will be captured on done page."),
+    )
+    gender_required = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Gender Required"),
+    )
+    activate_dob = models.BooleanField(
+        default=True,
+        editable=True,
+        verbose_name=_("Activate Date Of Birth"),
+    )
+    capture_dob_on_reg = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Capture On Registration"),
+        help_text=_("If Date Of Birth is activated, "
+                    "and Capture On Registration is not activated, "
+                    "The Date Of Birth field will be captured on done page."),
+    )
+    dob_required = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Date Of Birth Required"),
+    )
+    activate_location = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Activate Location"),
+    )
+    capture_location_on_reg = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Capture On Registration"),
+        help_text=_("If Location is activated, "
+                    "and Capture On Registration is not activated, "
+                    "The Location field will be captured on done page."),
+    )
+    location_required = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Location Required"),
+    )
+    activate_education_level = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Activate Education Level"),
+    )
+    capture_education_level_on_reg = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Capture On Registration"),
+        help_text=_("If Education Level is activated, "
+                    "and Capture On Registration is not activated, "
+                    "The Education Level field will be captured "
+                    "on done page."),
+    )
+    activate_education_level_required = models.BooleanField(
+        default=False,
+        editable=True,
+        verbose_name=_("Education Level Required"),
+    )
 
     panels = [
         MultiFieldPanel(
@@ -114,7 +205,42 @@ class UserProfilesSettings(BaseSetting):
             [
                 PageChooserPanel('terms_and_conditions'),
             ],
-            heading="Terms and Conditions on registration", )
+            heading="Terms and Conditions on registration", ),
+        MultiFieldPanel(
+            [
+                FieldPanel('activate_display_name'),
+                FieldPanel('capture_display_name_on_reg'),
+                FieldPanel('display_name_required'),
+            ],
+            heading="Display Name", ),
+        MultiFieldPanel(
+            [
+                FieldPanel('activate_gender'),
+                FieldPanel('capture_gender_on_reg'),
+                FieldPanel('gender_required'),
+            ],
+            heading="Gender", ),
+        MultiFieldPanel(
+            [
+                FieldPanel('activate_dob'),
+                FieldPanel('capture_dob_on_reg'),
+                FieldPanel('dob_required'),
+            ],
+            heading="Date Of Birth", ),
+        MultiFieldPanel(
+            [
+                FieldPanel('activate_location'),
+                FieldPanel('capture_location_on_reg'),
+                FieldPanel('location_required'),
+            ],
+            heading="Location", ),
+        MultiFieldPanel(
+            [
+                FieldPanel('activate_education_level'),
+                FieldPanel('capture_education_level_on_reg'),
+                FieldPanel('activate_education_level_required'),
+            ],
+            heading="Education Level", )
     ]
     # TODO: mobile_number_required field shouldn't be shown
     # if show_mobile_number_field is False
@@ -150,6 +276,18 @@ class UserProfile(models.Model):
     user = models.OneToOneField(User, related_name="profile", primary_key=True)
     date_of_birth = models.DateField(null=True)
     alias = models.CharField(
+        max_length=128,
+        blank=True,
+        null=True)
+    gender = models.CharField(
+        max_length=128,
+        blank=True,
+        null=True)
+    location = models.CharField(
+        max_length=128,
+        blank=True,
+        null=True)
+    education_level = models.CharField(
         max_length=128,
         blank=True,
         null=True)