+++ molo/profiles/views.py
@@ -28,9 +28,19 @@ class RegistrationView(FormView):
     def form_valid(self, form):
         username = form.cleaned_data["username"]
         password = form.cleaned_data["password"]
+        alias = form.cleaned_data["alias"]
+        date_of_birth = form.cleaned_data["date_of_birth"]
+        gender = form.cleaned_data["gender"]
+        location = form.cleaned_data["location"]
+        education_level = form.cleaned_data["education_level"]
         mobile_number = form.cleaned_data["mobile_number"]
         user = User.objects.create_user(username=username, password=password)
         user.profile.mobile_number = mobile_number
+        user.profile.alias = alias
+        user.profile.date_of_birth = date_of_birth
+        user.profile.gender = gender
+        user.profile.location = location
+        user.profile.education_level = education_level
         user.profile.site = self.request.site
         if form.cleaned_data["email"]:
             user.email = form.cleaned_data["email"]
@@ -63,15 +73,36 @@ class RegistrationView(FormView):
 
 
 class RegistrationDone(FormView):
-    """
-    Enables updating of the user's date of birth
-    """
-    form_class = forms.DateOfBirthForm
+    form_class = forms.DoneForm
     template_name = 'profiles/done.html'
 
     def form_valid(self, form):
         profile = self.request.user.profile
-        profile.date_of_birth = form.cleaned_data['date_of_birth']
+        if (UserProfilesSettings.for_site(
+            self.request.site).activate_dob) and not (
+            UserProfilesSettings.for_site(
+                self.request.site).capture_dob_on_reg):
+                profile.date_of_birth = form.cleaned_data["date_of_birth"]
+        if (UserProfilesSettings.for_site(
+            self.request.site).activate_display_name) and not (
+            UserProfilesSettings.for_site(
+                self.request.site).capture_display_name_on_reg):
+                profile.alias = form.cleaned_data["alias"]
+        if (UserProfilesSettings.for_site(
+            self.request.site).activate_gender) and not (
+            UserProfilesSettings.for_site(
+                self.request.site).capture_gender_on_reg):
+                profile.gender = form.cleaned_data["gender"]
+        if (UserProfilesSettings.for_site(
+            self.request.site).activate_location) and not (
+            UserProfilesSettings.for_site(
+                self.request.site).capture_location_on_reg):
+                profile.location = form.cleaned_data["location"]
+        if (UserProfilesSettings.for_site(
+            self.request.site).activate_education_level) and not (
+            UserProfilesSettings.for_site(
+                self.request.site).capture_education_level_on_reg):
+                profile.education_level = form.cleaned_data["education_level"]
         profile.save()
         return HttpResponseRedirect(form.cleaned_data.get('next', '/'))