+++ CHANGES.rst
@@ -14,6 +14,27 @@ CHANGE LOG
 ------------
 - Pre-release to support molo v5
 
+3.2.1
+-----
+- Add version number
+
+3.2.0
+-----
+- Add a success page after user log in
+
+3.1.1
+-----
+- Bug fixing on registration done page
+
+3.1.0
+-----
+- Add additional fields for registration
+- Display Name
+- Gender
+- Date of birth
+- Location
+- Education level
+
 3.0.0
 -----
 - Add terms and conditions on registration